Fix accidental deletion of unowned documents when replacing 'many'
Here's a way to accidentally destroy documents using MongoMapper:

  class Workgroup
    include MongoMapper::Document
    many :people
  end

  class Person
    include MongoMapper::Document
    belongs_to :workgroup
  end

  Person.create
  Person.create
  # Person.count is now 2.

  Workgroup.create(:people => [])
  # Person.count is now 0! Our two Person docs have been deleted.

What's going on here: The ':people => []' argument causes MongoMapper to
call #replace on a newly-created :people association.  This results in a
call to load_target, which eventually results results in a call to
scoped_conditions.

But scoped_conditions fails to correctly set up owner.id, and so
load_target queries the database for all Person documents that do not yet
have an owner.  Then, the replace function goes ahead and deletes all
these records.

So by writing:

  Workgroup.create(:people => [])

...MongoMapper ends up deleting all the Person objects that have not yet
been assigned to a Workgroup.  Needless to say, this is somewhat
surprising behavior.

This patch fixes two scoped_conditions methods to call
ensure_owner_saved, adds some error-checking to ensure_owner_saved, and
adds one final, last-ditch safety check to the 'replace' method to make
sure that we're actually deleting the correct documents.

I've tried to reproduce this error against the current master branch,
but the test case passes without any code changes.  However, examination
of the code suggests that this bug might still be present, but masked by
other changes in MongoMapper.

I'm making this patch against the currently released gem for users of
0.6.10 who want to avoid accidental data loss without upgrading to
'master'.